Add the category description
This will pick up the category description and make it available to the output of the module.

No effect on backwards compatibility, as it doesn't change what exists now, merely adds one more attribute.

Use Case: When you're choosing to output subcategories, it might be useful in some cases to have more than just the category name available. Output templates that don't need it can ignore it safely, but this makes it available for the templates that might want to use it.